[Hot Fix]

+++ 3D_Shooter/3D_Shooter/ShpereMesh.cpp
@@ -39,9 +39,9 @@ CShpereMesh::CShpereMesh()
 
 		for (int j = 0; j < sections; j++)
 		{
-			float x = cos(phi) * sin(theta);
-			float y = cos(theta);
-			float z = sin(phi) * sin(theta);
+			float x = (float)(cos(phi) * sin(theta));
+			float y = (float)(cos(theta));
+			float z = (float)(sin(phi) * sin(theta));
 
 			vertices[offset++] = x * radius;
 			vertices[offset++] = y * radius;
